+++ src/main/kotlin/com/chattriggers/ctjs/minecraft/libs/renderer/Rectangle.kt
@@ -5,7 +5,7 @@ import org.lwjgl.util.vector.Vector2f
 
 @External
 class Rectangle(
-        private var color: Int,
+        private var color: Long,
         private var x: Float,
         private var y: Float,
         private var width: Float,
@@ -14,8 +14,8 @@ class Rectangle(
     private var shadow = Shadow(this)
     private var outline = Outline(this)
 
-    fun getColor(): Int = this.color
-    fun setColor(color: Int) = apply { this.color = color }
+    fun getColor(): Long = this.color
+    fun setColor(color: Long) = apply { this.color = color }
 
     fun getX(): Float = this.x
     fun setX(x: Float) = apply { this.x = x }
@@ -42,10 +42,10 @@ class Rectangle(
     fun setShadowOffsetX(x: Float) = apply { this.shadow.offset.x = x }
     fun setShadowOffsetY(y: Float) = apply { this.shadow.offset.y = y }
 
-    fun getShadowColor(): Int = this.shadow.color
-    fun setShadowColor(color: Int) = apply { this.shadow.color = color }
+    fun getShadowColor(): Long = this.shadow.color
+    fun setShadowColor(color: Long) = apply { this.shadow.color = color }
 
-    fun setShadow(color: Int, x: Float, y: Float) = apply {
+    fun setShadow(color: Long, x: Float, y: Float) = apply {
         setShadow(true)
         setShadowColor(color)
         setShadowOffset(x, y)
@@ -54,13 +54,13 @@ class Rectangle(
     fun getOutline(): Boolean = this.outline.on
     fun setOutline(outline: Boolean) = apply { this.outline.on = outline }
 
-    fun getOutlineColor(): Int = this.outline.color
-    fun setOutlineColor(color: Int) = apply { this.outline.color = color }
+    fun getOutlineColor(): Long = this.outline.color
+    fun setOutlineColor(color: Long) = apply { this.outline.color = color }
 
     fun getThickness(): Float = this.outline.thickness
     fun setThickness(thickness: Float) = apply { this.outline.thickness = thickness }
 
-    fun setOutline(color: Int, thickness: Float) = apply {
+    fun setOutline(color: Long, thickness: Float) = apply {
         setOutline(true)
         setOutlineColor(color)
         setThickness(thickness)
@@ -75,7 +75,7 @@ class Rectangle(
     private class Shadow(
             val rect: Rectangle,
             var on: Boolean = false,
-            var color: Int = 0x50000000,
+            var color: Long = 0x50000000,
             var offset: Vector2f = Vector2f(5f, 5f)) {
         fun draw() {
             if (!this.on) return
@@ -97,7 +97,7 @@ class Rectangle(
     private class Outline(
             val rect: Rectangle,
             var on: Boolean = false,
-            var color: Int = 0xff000000.toInt(),
+            var color: Long = 0xff000000,
             var thickness: Float = 5f) {
         fun draw() {
             if (!this.on) return
